refactor: post/public/postList -> post/public/posts로 변경하고 컴포넌트 이동
