+++ src/features/post/public/posts/components/postList/PostListSkeleton.tsx
