+++ src/features/post/public/posts/store/PostSearchStateStore.tsx
